[Windows] Apply restructure
