+++ internal/collector/sysinfo/hardware/hardware_windows_test.go
@@ -1,4 +1,4 @@
-package sysinfo_test
+package hardware_test
 
 import (
 	"flag"
