+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/bad_memory_information
