+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/error_cpu_information
