+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/error_disk_information
