+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/error_gpu_information
