+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/error_memory_information
