+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/error_partition_information
