+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/error_product_information
