+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/error_screen_information
