+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/garbage_memory_information
