+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/malicious_disk_information
