+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/malicious_partition_information
