+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/missing_cpu_information
