+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/missing_disk_information
