+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/missing_gpu_information
