+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/missing_memory_information
