+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/missing_partition_information
