+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/missing_product_information
