+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/missing_screen_information
