+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/negative_memory_information
