+++ internal/collector/sysinfo/hardware/testdata/TestCollectWindows/golden/regular_hardware_information
